Recfa4 should be an integer...

+++ daily2longerfield.f90
@@ -9,7 +9,7 @@ program daily2longerfield
     include 'netcdf.inc'
     include 'getopts.inc'
     integer :: ntmax
-    real :: recfa4
+    integer :: recfa4
     parameter(ntmax=100000,recfa4=4)
     integer :: nperyear,nperyearnew,yr,mo,dy,i,j,n,itype,jx,jy
     integer :: nx,ny,nz,nt,firstyr,firstmo,lastyr,nvars, &
